+++ example/pubspec.yaml
@@ -12,7 +12,7 @@ dependencies:
   flutter:
     sdk: flutter
 
-  payhere_mobilesdk_flutter:
+  flutter_payhere:
     # When depending on this package from a real application you should use:
     #   payhere_mobilesdk_flutter: ^x.y.z
     # See https://dart.dev/tools/pub/dependencies#version-constraints
