+++ pubspec.yaml
@@ -1,16 +1,25 @@
-name: payhere_mobilesdk_flutter
+name: flutter_payhere
 description: Make PayHere payments with your Flutter App!
-version: 1.0.4
-homepage: https://github.com/PayHereLK/payhere-mobilesdk-flutter
+version: 1.0.0
+# homepage: https://github.com/PayHereLK/flutter_payhere
+publish_to: none
 
 environment:
-  # sdk: ">=2.7.0 <3.0.0"
   sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.20.0"
 
 dependencies:
   flutter:
     sdk: flutter
+  flutter_payhere_platform_interface:
+    git:
+      url: git://github.com/iwalpola/flutter_payhere.git
+      path: flutter_payhere_platform_interface
+  flutter_payhere_web:
+    git:
+      url: git://github.com/iwalpola/flutter_payhere.git
+      path: flutter_payhere_web
+  
 
 dev_dependencies:
   flutter_test:
@@ -24,3 +33,7 @@ flutter:
         pluginClass: PayhereMobilesdkFlutterPlugin
       ios:
         pluginClass: PayhereMobilesdkFlutterPlugin
+      web:
+        default_package: flutter_payhere_web
+# see following file for reference about organizing plugins
+# https://github.com/flutter/plugins/blob/master/packages/url_launcher/url_launcher/pubspec.yaml
